requirements.txt
pandas
yfinance
alpha_vantage
numpy
scipy
matplotlib
plotly
seaborn
arelle-release
pandas-datareader
cvxpy
python-dotenv
httpx
factor_analyzer
git+https://github.com/ayurrin/quantstats.git